+++ CHANGELOG.md
@@ -4,5 +4,5 @@ All notable changes to this project will be documented here.
 
 ## [0.1.0] - 2026-02-11
 ### Added
-- First release of the `snaplog` package featuring the `log` function and `SnapLogger` class
+- First release of the `snaplog` package featuring the `record`/`rec` and `log` functions and the `SnapLogger` class
 - Check back for more updates!